Ensure p-value adjustment applied to analysisSummary
