+++ packages/back-end/src/services/organizations.ts
@@ -88,6 +88,7 @@ import {
 import { createMetric } from "./experiments";
 import { isEmailEnabled, sendInviteEmail, sendNewMemberEmail } from "./email";
 import { ReqContextClass } from "./context";
+import { PValueCorrection } from "back-end/types/stats";
 
 export {
   getEnvironments,
@@ -197,6 +198,10 @@ export function getMetricDefaultsForOrg(context: ReqContext): MetricDefaults {
 
 export function getPValueThresholdForOrg(context: ReqContext): number {
   return context.org.settings?.pValueThreshold ?? DEFAULT_P_VALUE_THRESHOLD;
+}
+
+export function getPValueCorrectionForOrg(context: ReqContext): PValueCorrection {
+  return context.org.settings?.pValueCorrection ?? null;
 }
 
 export function getRole(
