+++ packages/front-end/components/Experiment/BreakDownResults.tsx
@@ -13,6 +13,8 @@ import {
 import {
   expandMetricGroups,
   ExperimentMetricInterface,
+  setAdjustedCIs,
+  setAdjustedPValuesOnResults,
 } from "shared/experiments";
 import { isDefined } from "shared/util";
 import { FaAngleRight, FaUsers } from "react-icons/fa";
@@ -20,9 +22,7 @@ import Collapsible from "react-collapsible";
 import { useDefinitions } from "@/services/DefinitionsContext";
 import {
   applyMetricOverrides,
-  setAdjustedPValuesOnResults,
   ExperimentTableRow,
-  setAdjustedCIs,
   hasRisk,
 } from "@/services/experiments";
 import ResultsTable from "@/components/Experiment/ResultsTable";
