+++ packages/front-end/components/Experiment/CompactResults.tsx
@@ -22,14 +22,14 @@ import {
   expandMetricGroups,
   ExperimentMetricInterface,
   getMetricLink,
+  setAdjustedCIs,
+  setAdjustedPValuesOnResults,
 } from "shared/experiments";
 import { isDefined } from "shared/util";
 import { useDefinitions } from "@/services/DefinitionsContext";
 import {
   applyMetricOverrides,
-  setAdjustedPValuesOnResults,
   ExperimentTableRow,
-  setAdjustedCIs,
   hasRisk,
 } from "@/services/experiments";
 import { GBCuped } from "@/components/Icons";
