+++ packages/front-end/services/experiments.ts
@@ -1,12 +1,9 @@
 import { SnapshotMetric } from "back-end/types/experiment-snapshot";
 import {
   DifferenceType,
-  PValueCorrection,
   StatsEngine,
 } from "back-end/types/stats";
-import normal from "@stdlib/stats/base/dists/normal";
 import {
-  ExperimentReportResultDimension,
   ExperimentReportVariationWithIndex,
   MetricSnapshotSettings,
 } from "back-end/types/report";
@@ -240,170 +237,6 @@ export function pValueFormatter(pValue: number, digits: number = 3): string {
   return pValue < Math.pow(10, -digits)
     ? `<0.${"0".repeat(digits - 1)}1`
     : pValue.toFixed(digits);
-}
-
-export type IndexedPValue = {
-  pValue: number;
-  index: (number | string)[];
-};
-
-export function adjustPValuesBenjaminiHochberg(
-  indexedPValues: IndexedPValue[]
-): IndexedPValue[] {
-  const newIndexedPValues = cloneDeep<IndexedPValue[]>(indexedPValues);
-  const m = newIndexedPValues.length;
-
-  newIndexedPValues.sort((a, b) => {
-    return b.pValue - a.pValue;
-  });
-  newIndexedPValues.forEach((p, i) => {
-    newIndexedPValues[i].pValue = Math.min((p.pValue * m) / (m - i), 1);
-  });
-
-  let tempval = newIndexedPValues[0].pValue;
-  for (let i = 1; i < m; i++) {
-    if (newIndexedPValues[i].pValue < tempval) {
-      tempval = newIndexedPValues[i].pValue;
-    } else {
-      newIndexedPValues[i].pValue = tempval;
-    }
-  }
-  return newIndexedPValues;
-}
-
-export function adjustPValuesHolmBonferroni(
-  indexedPValues: IndexedPValue[]
-): IndexedPValue[] {
-  const newIndexedPValues = cloneDeep<IndexedPValue[]>(indexedPValues);
-  const m = newIndexedPValues.length;
-  newIndexedPValues.sort((a, b) => {
-    return a.pValue - b.pValue;
-  });
-  newIndexedPValues.forEach((p, i) => {
-    newIndexedPValues[i].pValue = Math.min(p.pValue * (m - i), 1);
-  });
-
-  let tempval = newIndexedPValues[0].pValue;
-  for (let i = 1; i < m; i++) {
-    if (newIndexedPValues[i].pValue > tempval) {
-      tempval = newIndexedPValues[i].pValue;
-    } else {
-      newIndexedPValues[i].pValue = tempval;
-    }
-  }
-  return newIndexedPValues;
-}
-
-export function setAdjustedPValuesOnResults(
-  results: ExperimentReportResultDimension[],
-  nonGuardrailMetrics: string[],
-  adjustment: PValueCorrection
-): void {
-  if (!adjustment) {
-    return;
-  }
-
-  let indexedPValues: IndexedPValue[] = [];
-  results.forEach((r, i) => {
-    r.variations.forEach((v, j) => {
-      nonGuardrailMetrics.forEach((m) => {
-        const pValue = v.metrics[m]?.pValue;
-        if (pValue !== undefined) {
-          indexedPValues.push({
-            pValue: pValue,
-            index: [i, j, m],
-          });
-        }
-      });
-    });
-  });
-
-  if (indexedPValues.length === 0) {
-    return;
-  }
-
-  if (adjustment === "benjamini-hochberg") {
-    indexedPValues = adjustPValuesBenjaminiHochberg(indexedPValues);
-  } else if (adjustment === "holm-bonferroni") {
-    indexedPValues = adjustPValuesHolmBonferroni(indexedPValues);
-  }
-
-  // modify results in place
-  indexedPValues.forEach((ip) => {
-    const ijk = ip.index;
-    results[ijk[0]].variations[ijk[1]].metrics[ijk[2]].pValueAdjusted =
-      ip.pValue;
-  });
-  return;
-}
-
-export function adjustedCI(
-  adjustedPValue: number,
-  lift: number | undefined,
-  pValueThreshold: number
-): [number, number] {
-  if (!lift) return [0, 0];
-  const zScore = normal.quantile(1 - pValueThreshold / 2, 0, 1);
-  const adjStdDev = Math.abs(
-    lift / normal.quantile(1 - adjustedPValue / 2, 0, 1)
-  );
-  const width = zScore * adjStdDev;
-  return [lift - width, lift + width];
-}
-
-export function setAdjustedCIs(
-  results: ExperimentReportResultDimension[],
-  pValueThreshold: number
-): void {
-  results.forEach((r) => {
-    r.variations.forEach((v) => {
-      for (const key in v.metrics) {
-        const pValueAdjusted = v.metrics[key].pValueAdjusted;
-        const uplift = v.metrics[key].uplift;
-        const ci = v.metrics[key].ci;
-        if (
-          pValueAdjusted === undefined ||
-          uplift === undefined ||
-          ci === undefined
-        ) {
-          continue;
-        }
-
-        const adjCI = getAdjustedCI(
-          pValueAdjusted,
-          uplift.mean,
-          pValueThreshold,
-          ci
-        );
-        if (adjCI) {
-          v.metrics[key].ciAdjusted = adjCI;
-        } else {
-          v.metrics[key].ciAdjusted = ci;
-        }
-      }
-    });
-  });
-  return;
-}
-
-export function getAdjustedCI(
-  pValueAdjusted: number,
-  lift: number | undefined,
-  pValueThreshold: number,
-  ci: [number, number]
-): [number, number] | undefined {
-  // set to Inf if adjusted pValue is 1
-  if (pValueAdjusted > 0.999999) {
-    return [-Infinity, Infinity];
-  }
-
-  const adjCI = adjustedCI(pValueAdjusted, lift, pValueThreshold);
-  // only update if CI got wider, should never get more narrow
-  if (adjCI[0] < ci[0] && adjCI[1] > ci[1]) {
-    return adjCI;
-  } else {
-    return ci;
-  }
 }
 
 export type RowResults = {
